Enhancement: Versioning Endpoints
Currently if the backend changes, the changes also have to be applied in the Frontend. This is
not the case if the endpoints are versioned. Changing the backend should
result in a new endpoint : /v2/ ... Then the frontend can still work
with v1 endpoint and update when needed

BREAKING CHANGES: The old endpoints /UUID and /ws are no longer reachable
with this new version