Refactor Repository
This commit adds a `.gitignore` to make sure python temporaries are not
submitted.

Furthermore it updates the readme texts, by adding more context and some
elaborations. Requesting the websocket was moved to the top-level readme
in order to ensure a mutual understanding, that OmnAIView-DevDataServer
written in other languages use the same interface when it comes to
the creation of a datastream.

To enhance usability, a requirements.txt was added to the python
example.